+++ index.html
@@ -11,8 +11,10 @@
     <link rel="stylesheet" href="./style.css">
     <link href="https://fonts.googleapis.com/css?family=Open+Sans&display=swap" rel="stylesheet">
     <script type="module" src="/js/physi.js"></script>
+    <script type="module" src="/js/physijs_worker.js"></script>
     <script type="module" src="/js/three.js"></script>
     <script type="module" src="/scripts/audio.js"></script>
+    <script type="module" src="/scripts/geometry.js"></script>
 
 
   </head>
